requirements.txt
Flask==3.0.3
tensorflow==2.17.0
pillow==11.0.0